:sparkles: Implemented Initial Screen
