--- Challange HURB - Alpha/Challange HURB - Alpha/Controller/ViewController.swift
@@ -1,26 +1,0 @@
-//
-//  ViewController.swift
-//  Challange HURB - Alpha
-//
-//  Created by Luiz Fernando Cunha Duarte on 26/10/19.
-//  Copyright © 2019 Luiz Fernando Cunha Duarte. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController, StoryboardInitializable {
-    weak var coordinator: MainCoordinator?
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        
-        print("Apareceu")
-    }
-
-}
-
