+++ Challange HURB - Alpha/Challange HURB - Alpha/Coordinator/MainCoordinator.swift
@@ -17,10 +17,11 @@ class MainCoordinator: Coordinator {
     
     init(navigationController: UINavigationController) {
         self.navigationController = navigationController
+        self.navigationController.isNavigationBarHidden = true
     }
     
     func start() {
-        let startViewController = ViewController.instantiate()
+        let startViewController = InitialViewController.instantiate()
         startViewController.coordinator = self
         navigationController.pushViewController(startViewController, animated: false)
     }
